polystat-backend/requirements.txt
fastapi==0.116.1
pydantic==2.11.7
pydantic-settings==2.10.1
uvicorn==0.35.0
python-dotenv==1.1.1
httpx==0.28.1
